Create basic interfaces and real-valued implementations for constants and variables
Create the utility interfaces Immutable and CachedHashCode.

Create the base interfaces Expression and its sub-interfaces Constant and Variable, which will serve as base interfaces
for expressions of all types. If deemed useful, some of these may be turned into abstract base classes in the future.

Create abstract base class RealExpression to serve as basis for all real-value expressions, along with the
implementations RealVariable and RealConstant for real-valued terminal expressions.

Create class RealAssignment which maps RealVariable instances to double values, and can be used to evaluate
RealExpressions, e.g. in visitors.