Create abstract class for non-terminal expressions and implementation for negation
Abstract class RealSuperExpression provides a common implementation of aggregating contained variables from
sub-expressions, so all non-terminal expression implementations can simply pass their list of sub-expressions to the
protected constructor in this class.

An abstract subclass of this, RealExpressionUnary, serves as a base class for all unary non-terminal expressions, such
as negation, providing a constructor that takes a single sub-expression instead of a list of them.

Also create concrete implementation of a unary super-expression, RealExprNegation, including a static factory method.