Add options to sort subexpression lists of non-unary super expressions
Subexpressions are represented as lists for two reasons:
- the same subexpression (instance) may appear more than once per super-expression, so a simple Set wouldn't suffice
  and we would need some kind of multi-set instead; since no multi-sets are provided by the Java standard library, we
  prefer to work without one for the moment; and
- while the currently existing non-unary super-expressions (addition and multiplication) are commutative, not all super
  expressions need to be so; using an order-preserving list leaves the API ready to handle non-commutative expression
  types.

As such, since the order of elements factors in the hash code (and equality) of Lists, and therefore of
RealSuperExpressions, we need to normalize (i.e. sort) subexpression lists for commutative operations.

We continue to offer constructor (and factory method) variants that do *not* sort the subexpression list, to allow code
where it is contextually ensured that the list is already sorted a more efficient entry point.